refactor: 인수 테스트 클래스들의 static 메서드 상위 이동 및 테스트 유틸 삭제
