+++ docs/readme.md
@@ -1,6 +1,7 @@
 # 구현할 기능 목록 step4
 
 ## 요금항목 추가
+
 - [x] LineRequest dto
 - [x] Line
 - [x] DB Line table
@@ -10,9 +11,9 @@
 - [x] Fare 에 추가 노선 요금 적용
 - [x] Line 인수테스트 수정
 - [x] Fare 도메인테스트 수정
-    - [x] Age 적용 도메인 테스트 
+    - [x] Age 적용 도메인 테스트
 
 ## 로그인
 
-- [ ] 경로조회 관련 로그인 인수테스트 작성
+- [x] 경로조회 관련 로그인 인수테스트 작성
