+++ src/test/java/wooteco/subway/AcceptanceTest.java
@@ -1,11 +1,25 @@
 package wooteco.subway;
 
 import io.restassured.RestAssured;
+import io.restassured.response.ExtractableResponse;
+import io.restassured.response.Response;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ActiveProfiles;
+import wooteco.subway.auth.dto.TokenResponse;
+import wooteco.subway.line.dto.LineRequest;
+import wooteco.subway.line.dto.SectionRequest;
+import wooteco.subway.member.dto.MemberRequest;
+import wooteco.subway.station.dto.StationRequest;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
@@ -14,6 +28,88 @@ public class AcceptanceTest {
 
     @LocalServerPort
     private int port;
+
+    public static ExtractableResponse<Response> registerMember(String email, String password, int age) {
+        MemberRequest memberRequest = new MemberRequest(email, password, age);
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(memberRequest)
+                .when().post("/members")
+                .then().log().all()
+                .extract();
+    }
+
+    public static TokenResponse login(String email, String password) {
+        Map<String, String> params = new HashMap<>();
+        params.put("email", email);
+        params.put("password", password);
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(params)
+                .when().post("/login/token")
+                .then().log().all()
+                .extract()
+                .as(TokenResponse.class);
+    }
+
+
+    public static ExtractableResponse<Response> getMyMemberInfo(String token) {
+        return RestAssured.given().log().all()
+                .auth().oauth2(token)
+                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .when().get("/members/me")
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> createStation(String name, String token) {
+        StationRequest stationRequest = new StationRequest(name);
+        return RestAssured
+                .given().log().all()
+                .body(stationRequest)
+                .auth().oauth2(token)
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .when().post("/stations")
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> createLine(LineRequest params, String token) {
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .auth().oauth2(token)
+                .body(params)
+                .when().post("/lines")
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> addSection(Long id, SectionRequest sectionRequest, String loginToken) {
+        return RestAssured.given().log().all()
+                .contentType(MediaType.APPLICATION_JSON_VALUE)
+                .body(sectionRequest)
+                .auth().oauth2(loginToken)
+                .when().post("/lines/{lineId}/sections", id)
+                .then().log().all()
+                .extract();
+    }
+
+    public static ExtractableResponse<Response> findOneLine(long id, String token) {
+        return RestAssured.given().log().all()
+                .auth().oauth2(token)
+                .accept(MediaType.APPLICATION_JSON_VALUE)
+                .when().get("/lines/" + id)
+                .then().log().all()
+                .extract();
+    }
+
+    public static void assertResponseStatus(ExtractableResponse<Response> response, HttpStatus httpStatus) {
+        assertThat(response.statusCode()).isEqualTo(httpStatus.value());
+    }
+
+    public static void assertResponseMessage(ExtractableResponse<Response> response, String message) {
+        assertThat(response.body().asString()).isEqualTo(message);
+    }
 
     @BeforeEach
     public void setUp() {
